Use consistent type for vreplication id (#12218)
* Use consistent type for vreplication id

The `_vt.vreplication` table has an `INT` type without `UNSIGNED`, so it
maps to the Go `int32` type to have the exact same consistent range.

This means that we should use `int32` consistently for everywhere that
we use the VReplication identifier.

We had int32, uint32 and int used pretty randomly for these which leads
to CodeQL flagging these and how this was found.

Instead this now converts everything to int32. It's also safe to change
the protobuf, since the protobuf wire format for int32, uint32, int64 &
uint64 is all compatible and actually uses the same variable int length
format.

The only exception would be if someone has a VReplication identifier
that would not fit in `int32`, but well, this PR shows that in that case
we already blow up in unexpected and unknown ways anyway. So it's better
to have that explicitly error.


* Add support for all integer types for bind variables

This ensures we cover all the Go int types to generate bind parameters
from.


* Regenerate protos for vtadmin


* Fix test for using ParseInt


---------

+++ go/vt/vttablet/onlineddl/vrepl.go
@@ -46,7 +46,7 @@ import (
 
 // VReplStream represents a row in _vt.vreplication table
 type VReplStream struct {
-	id                   int64
+	id                   int32
 	workflow             string
 	source               string
 	pos                  string
